Updating vin endpoint
